refactor: util関数を整理し、ファイル構成を改善
feat: 新しいキャッシュ機能を追加し、変数の取得を最適化
feat: テキストノードをフィルタリングする関数を追加
feat: 先祖インスタンスを取得する関数を追加
feat: ローカル変数を取得する関数を追加
feat: ライブラリ変数を取得する関数を追加
fix: エラーハンドリングを改善
