+++ src/main/utils/cache.ts
