+++ src/main/utils/getVariable.ts
